requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
pinecone-client
tiktoken
pypdf